+++ crates/client/rpc/src/methods/write/add_invoke_transaction.rs
@@ -1,15 +1,9 @@
 use jsonrpsee::core::RpcResult;
-use mc_sync::utility::{chain_id, feeder_gateway, gateway};
-use mp_hashers::HasherT;
-use mp_types::block::DBlockT;
-use sc_client_api::backend::{Backend, StorageProvider};
-use sc_client_api::BlockBackend;
-use sp_blockchain::HeaderBackend;
 use starknet_core::types::{BroadcastedInvokeTransaction, InvokeTransactionResult};
-use starknet_providers::{Provider, ProviderError, SequencerGatewayProvider};
+use starknet_providers::{Provider, ProviderError};
 
 use crate::errors::StarknetRpcApiError;
-use crate::Starknet;
+use crate::{bail_internal_server_error, Starknet};
 
 /// Add an Invoke Transaction to invoke a contract function
 ///
@@ -20,26 +14,18 @@ use crate::Starknet;
 /// # Returns
 ///
 /// * `transaction_hash` - transaction hash corresponding to the invocation
-pub async fn add_invoke_transaction<BE, C, H>(
-    _starknet: &Starknet<BE, C, H>,
+pub async fn add_invoke_transaction(
+    starknet: &Starknet,
     invoke_transaction: BroadcastedInvokeTransaction,
-) -> RpcResult<InvokeTransactionResult>
-where
-    BE: Backend<DBlockT> + 'static,
-    C: HeaderBackend<DBlockT> + BlockBackend<DBlockT> + StorageProvider<DBlockT, BE> + 'static,
-    H: HasherT + Send + Sync + 'static,
-{
-    let sequencer = SequencerGatewayProvider::new(feeder_gateway(), gateway(), chain_id());
+) -> RpcResult<InvokeTransactionResult> {
+    let sequencer = starknet.make_sequencer_provider();
 
     let sequencer_response = match sequencer.add_invoke_transaction(invoke_transaction).await {
         Ok(response) => response,
         Err(ProviderError::StarknetError(e)) => {
             return Err(StarknetRpcApiError::from(e).into());
         }
-        Err(e) => {
-            log::error!("Failed to add invoke transaction to sequencer: {e}");
-            return Err(StarknetRpcApiError::InternalServerError.into());
-        }
+        Err(e) => bail_internal_server_error!("Failed to add invoke transaction to sequencer: {e}"),
     };
 
     Ok(sequencer_response)