+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - fix(hashers): cleaned hashers using types core hashers and Felt
+- refactor: remove substrate block storage
 - feat(infra): Added boilerplate to deploy a grafana/prometheus dashboard
 - refacor: use db hash
 - refactor: l2-sync